Extract terminal private-mode tracking from HeadlessEmulator
Merging main's Unicode11 mirror work with this branch's OSC7 path-flavor
options pushed headless-emulator.ts past the max-lines cap. The DECSET
mouse/SGR scanning is a self-contained state machine, so it moves to
terminal-private-mode-tracker.ts unchanged.
