+++ src/main/daemon/headless-emulator.ts
@@ -7,6 +7,7 @@ import { extractLastOscTitle } from '../../shared/agent-detection'
 import { collectHeadlessOscLinkRanges } from './headless-osc-link-ranges'
 import { extractOscScanTail, scanOsc7Uris } from './osc7-uri-extraction'
 import { parseFileUriPath } from './osc7-file-uri'
+import { TerminalPrivateModeTracker } from './terminal-private-mode-tracker'
 import type { TerminalSnapshot, TerminalModes } from './types'
 import type { TerminalOscLinkRange } from '../../shared/terminal-osc-link-ranges'
 
@@ -26,9 +27,6 @@ type TerminalWithSynchronousWrite = Terminal & {
 
 const DEFAULT_SCROLLBACK = 5000
 const OSC_SCAN_TAIL_LIMIT = 4096
-// Why: PTY/SSH chunks can split a long combined DECSET before the final h/l.
-// Keep parser state far beyond normal mode lists while still bounding memory.
-const PRIVATE_MODE_SCAN_TAIL_LIMIT = 4096
 
 export class HeadlessEmulator {
   private terminal: Terminal
@@ -36,10 +34,7 @@ export class HeadlessEmulator {
   private cwd: string | null = null
   private lastTitle: string | null = null
   private oscScanTail = ''
-  private privateModeScanTail = ''
-  private mouseTrackingMode: NonNullable<TerminalModes['mouseTrackingMode']> = 'none'
-  private sgrMouseMode = false
-  private sgrMousePixelsMode = false
+  private privateModes = new TerminalPrivateModeTracker()
   private restoredOscLinks: TerminalOscLinkRange[] = []
   private disposed = false
   private readonly pathFlavor?: 'posix' | 'win32'
@@ -93,7 +88,7 @@ export class HeadlessEmulator {
       this.terminal.write(data, () => {
         // Why: snapshots combine serialized xterm state with mirrored mouse
         // modes. Commit the mirror only after xterm has parsed the same bytes.
-        this.scanPrivateModes(data)
+        this.privateModes.scan(data)
         resolve()
       })
     })
@@ -119,7 +114,7 @@ export class HeadlessEmulator {
     // Why: hidden renderer restore snapshots are requested immediately after
     // PTY bursts; queued headless writes can snapshot half-cleared TUI rows.
     writeSync.call((this.terminal as TerminalWithSynchronousWrite)._core, data)
-    this.scanPrivateModes(data)
+    this.privateModes.scan(data)
     return true
   }
 
@@ -222,78 +217,6 @@ export class HeadlessEmulator {
     return extractOscScanTail(input, OSC_SCAN_TAIL_LIMIT)
   }
 
-  private scanPrivateModes(data: string): void {
-    const input = this.privateModeScanTail + data
-    this.privateModeScanTail = this.extractPrivateModeScanTail(input)
-    // oxlint-disable-next-line no-control-regex -- terminal escape sequences require control chars
-    const privateModeRe = /\x1bc|\x1b\[\?([0-9;]+)([hl])|\x9b\?([0-9;]+)([hl])/g
-    let match: RegExpExecArray | null
-    while ((match = privateModeRe.exec(input)) !== null) {
-      if (match[0] === '\x1bc') {
-        this.mouseTrackingMode = 'none'
-        this.sgrMouseMode = false
-        this.sgrMousePixelsMode = false
-        continue
-      }
-      const params = match[1] ?? match[3]
-      const enabled = (match[2] ?? match[4]) === 'h'
-      for (const rawParam of params.split(';')) {
-        if (rawParam === '') {
-          continue
-        }
-        const param = Number(rawParam)
-        if (!Number.isInteger(param)) {
-          continue
-        }
-        if (param === 9) {
-          this.mouseTrackingMode = enabled ? 'x10' : 'none'
-        }
-        if (param === 1000) {
-          this.mouseTrackingMode = enabled ? 'vt200' : 'none'
-        }
-        if (param === 1002) {
-          this.mouseTrackingMode = enabled ? 'drag' : 'none'
-        }
-        if (param === 1003) {
-          this.mouseTrackingMode = enabled ? 'any' : 'none'
-        }
-        if (param === 1006) {
-          this.sgrMouseMode = enabled
-          this.sgrMousePixelsMode = false
-        }
-        if (param === 1016) {
-          this.sgrMouseMode = false
-          this.sgrMousePixelsMode = enabled
-        }
-      }
-    }
-  }
-
-  private extractPrivateModeScanTail(input: string): string {
-    const start = Math.max(input.lastIndexOf('\x1b'), input.lastIndexOf('\x9b'))
-    if (start === -1) {
-      return ''
-    }
-    const tail = input.slice(start)
-    if (tail.length > PRIVATE_MODE_SCAN_TAIL_LIMIT) {
-      return ''
-    }
-    if (tail === '\x1b' || tail === '\x1b[' || tail === '\x9b') {
-      return tail
-    }
-    if (tail.startsWith('\x1b[?')) {
-      return this.isIncompletePrivateModeParams(tail.slice(3)) ? tail : ''
-    }
-    if (tail.startsWith('\x9b?')) {
-      return this.isIncompletePrivateModeParams(tail.slice(2)) ? tail : ''
-    }
-    return ''
-  }
-
-  private isIncompletePrivateModeParams(params: string): boolean {
-    return /^[0-9;]*$/.test(params)
-  }
-
   private normalizeSnapshotAnsiForModes(snapshotAnsi: string, modes: TerminalModes): string {
     if (!modes.alternateScreen) {
       return snapshotAnsi
@@ -321,13 +244,13 @@ export class HeadlessEmulator {
 
   private getModes(): TerminalModes {
     const buffer = this.terminal.buffer.active
-    const mouseTrackingMode = this.mouseTrackingMode
+    const mouseTrackingMode = this.privateModes.mouseTrackingMode
     return {
       bracketedPaste: this.terminal.modes.bracketedPasteMode,
       mouseTracking: mouseTrackingMode !== 'none',
       mouseTrackingMode,
-      sgrMouseMode: this.sgrMouseMode,
-      sgrMousePixelsMode: this.sgrMousePixelsMode,
+      sgrMouseMode: this.privateModes.sgrMouseMode,
+      sgrMousePixelsMode: this.privateModes.sgrMousePixelsMode,
       applicationCursor:
         buffer.type === 'normal' ? this.terminal.modes.applicationCursorKeysMode : false,
       alternateScreen: buffer.type === 'alternate'
